+++ src/Permission.php
@@ -6,7 +6,7 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\rbac;
+namespace Yiisoft\Rbac;
 
 /**
  * For more details and usage information on Permission, see the [guide article on security authorization](guide:security-authorization).
