+++ src/Role.php
@@ -6,7 +6,7 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\rbac;
+namespace Yiisoft\Rbac;
 
 /**
  * For more details and usage information on Role, see the [guide article on security authorization](guide:security-authorization).
