+++ src/RuleFactoryInterface.php
@@ -6,7 +6,7 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\rbac;
+namespace Yiisoft\Rbac;
 
 /**
  * @author Andrii Vasyliev <sol@hiqdev.com>
