+++ src/exceptions/ExceptionInterface.php
@@ -6,7 +6,7 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\rbac\exceptions;
+namespace Yiisoft\Rbac\Exceptions;
 
 /**
  * ExceptionInterface to mark exceptions in this package.
