+++ tests/PgSQLManagerTest.php
@@ -6,24 +6,24 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\rbac\tests\unit;
+namespace Yiisoft\Rbac\Tests;
 
-use yii\rbac\DbManager;
-use yii\rbac\DIRuleFactory;
+use Yiisoft\Rbac\DbManager;
+use Yiisoft\Rbac\DIRuleFactory;
 
 /**
- * MySQLManagerTest.
+ * PgSQLManagerTest.
  *
  * @group db
  * @group rbac
- * @group mysql
+ * @group pgsql
  */
-class MySQLManagerTest extends DbManagerTestCase
+class PgSQLManagerTest extends DbManagerTestCase
 {
-    protected static $driverName = 'mysql';
+    protected static $driverName = 'pgsql';
 
     /**
-     * @return \yii\rbac\ManagerInterface
+     * @return \Yiisoft\Rbac\ManagerInterface
      */
     protected function createManager()
     {
